requirements.txt
scipy==1.6.3
numpy==1.19.5
matplotlib==3.4.2
pygame==2.0.1
scikit_image==0.18.2
opencv_python==4.5.2.52
Pillow==8.3.0
skimage==0.0
